Add support for throwables to the Kit Profiler
This required changing up a few things. I had to make the Tools tab
wider, and the Kit Profiler tabs wider as well. I had to add supportfor
forcibly truncating text in cells when drawing rows of text on the
screen. I had to setup an align table for drawing the rows of configured
kits.

I then added the necessary getters and setters and equippers for
throwables. It all appears to work just fine. This required updatinga
translation string, which I've done for all languages supported by
PocoHud as best as possible.

I also removed the spaces between entries in the translation string I
touched, and updated the translation for "MELEE" for a few languages.